Update _blast.py
Here is my documentation for why issue #155(https://github.com/qiime2/q2-feature-classifier/issues/155) is not accomplishable at this time. Issue #77 (https://github.com/qiime2/q2-feature-classifier/issues/77) was fixed by removing -num_threads as it was ignored when subject was specified. Since num_threads is currently not a variable it can not be exposed.